+++ components/render/opengl_driver/sources/render_target_manager/fbo_manager/gl_fbo_frame_buffer_manager.cpp
@@ -69,7 +69,7 @@ IFrameBuffer* FboFrameBufferManager::CreateFrameBuffer (View* color_view, View* 
     Установка текущего буфера кадра
 */
 
-void FboFrameBufferManager::SetFrameBuffer (size_t fbo_id, size_t cache_id)
+void FboFrameBufferManager::SetFrameBuffer (unsigned int fbo_id, size_t cache_id)
 {
   static const char* METHOD_NAME = "render::low_level::opengl::FboFrameBufferManager::SetFrameBuffer";
