+++ components/render/opengl_driver/sources/render_target_manager/fbo_manager/gl_fbo_render_buffer.cpp
@@ -172,7 +172,7 @@ FboRenderBuffer::~FboRenderBuffer ()
     Получение OpenGL идентификатора буфера кадра для возможности ручной записи / чтения данных
 */
 
-size_t FboRenderBuffer::GetFrameBufferId ()
+unsigned int FboRenderBuffer::GetFrameBufferId ()
 {
   static const char* METHOD_NAME = "render::low_level::opengl::FboRenderBuffer::GetFrameBufferId";
