+++ components/render/opengl_driver/sources/shader_stage/ffp/gl_ffp_shader_manager.cpp
@@ -47,7 +47,7 @@ IShader* FfpShaderManager::CreateShader  (const ShaderDesc& shader_desc, const L
     Создание программы
 */
 
-ICompiledProgram* FfpShaderManager::CreateProgram (size_t shaders_count, IShader** shaders, const LogFunction&)
+ICompiledProgram* FfpShaderManager::CreateProgram (unsigned int shaders_count, IShader** shaders, const LogFunction&)
 {
   static const char* METHOD_NAME = "render::low_level::opengl::FfpShaderManager::CreateProgram";
